fixes2; better components logic
