+++ types/NavItems.ts
@@ -1,0 +1,11 @@
+interface TextNavItem {
+  linkSlug: string;
+  content: string;
+}
+
+interface IconNavItem {
+  linkSlug: string;
+  iconName: string;
+}
+
+export type { TextNavItem, IconNavItem };
